# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Capital-Challenge backend API.",
        "title": "Capital-Challenge API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/v1",
    "paths": {
        "/companies/{ticker}": {
            "get": {
                "description": "get company info by Ticker",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "companies"
                ],
                "summary": "Get Company info",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Ticker",
                        "name": "ticker",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1Companies.GetCompanyInfoResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/company-stocks": {
            "get": {
                "description": "get the daily information about company stocks",
                "tags": [
                    "company_stock"
                ],
                "summary": "Get Daily Company Stock",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/company-stocks/buy/{user_id}": {
            "post": {
                "description": "buy the selected amount of company stock",
                "tags": [
                    "company_stock"
                ],
                "summary": "BuyCompanyStock",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user_id",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1companystock.BuyCompanyStockRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.UserTransactions"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/company-stocks/sell/{user_id}": {
            "post": {
                "description": "sell the selected amount of company stock",
                "tags": [
                    "company_stock"
                ],
                "summary": "SellCompanyStock",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user_id",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1companystock.SellCompanyStockRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.UserTransactions"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/company-stocks/{user_id}": {
            "post": {
                "description": "buy or sell the selected amount of company stock",
                "tags": [
                    "company_stock"
                ],
                "summary": "BuyOrSellCompanyStock",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user_id",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1companystock.BuyOrSellCompanyStocksRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.UserTransactions"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.Companies": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "currency_name": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "homepage_url": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "locale": {
                    "type": "string"
                },
                "logo_url": {
                    "type": "string"
                },
                "market": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "primary_exchange": {
                    "type": "string"
                },
                "ticker": {
                    "type": "string"
                },
                "total_employees": {
                    "type": "number"
                },
                "type": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.CompanyStock": {
            "type": "object",
            "properties": {
                "close_price": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "highest_price": {
                    "type": "number"
                },
                "id": {
                    "type": "string"
                },
                "lowest_price": {
                    "type": "number"
                },
                "number_of_transactions": {
                    "type": "integer"
                },
                "open_price": {
                    "type": "number"
                },
                "otc": {
                    "type": "boolean"
                },
                "ticker": {
                    "type": "string"
                },
                "trading_volume": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "volume_weighted_average_price": {
                    "type": "number"
                }
            }
        },
...
        "models.UserTransactions": {
            "type": "object",
            "properties": {
                "buy_price": {
                    "type": "number"
                },
                "company_stock_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                },
                "sell_price": {
                    "type": "number"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "v1Companies.GetCompanyInfoResponse": {
            "type": "object",
            "properties": {
                "company": {
                    "$ref": "#/definitions/models.Companies"
                },
                "company_stock": {
                    "$ref": "#/definitions/models.CompanyStock"
                }
            }
        },
...
        "v1companystock.BuyCompanyStockRequest": {
            "type": "object",
            "properties": {
                "company_stock_id": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                }
            }
        },
        "v1companystock.BuyOrSellCompanyStocksRequest": {
            "type": "object",
            "properties": {
                "buy_or_sell": {
                    "type": "string"
                },
                "company_stock_id": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                }
            }
        },
        "v1companystock.SellCompanyStockRequest": {
            "type": "object",
            "properties": {
                "company_stock_id": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}